requirements.txt
requests
Pillow
schedule
selenium
